refactor(primitives): move DID and SigningKeyId from scp-identity to scp-primitives (#1446)
Prerequisite for scp-protocol crate extraction. DID and SigningKeyId are
pure types with zero async dependencies — they belong in the leaf crate.
scp-identity re-exports for backward compatibility. scp-event-log drops
its scp-identity dependency, breaking the tokio transitive chain.
