+++ crates/scp-identity/Cargo.toml
@@ -11,6 +11,7 @@ keywords = ["did", "identity", "decentralized", "dht"]
 categories = ["authentication", "cryptography"]
 
 [dependencies]
+scp-primitives = { path = "../scp-primitives", version = "=0.1.0-beta.1" }
 scp-platform = { path = "../scp-platform", version = "=0.1.0-beta.1" }
 tokio = { workspace = true }
 serde = { workspace = true }
