+++ crates/scp-identity/src/lib.rs
@@ -52,7 +52,7 @@ pub use dht::{
     did_from_ed25519_public_key, extract_public_key, verify_bep44_signature, verify_migration,
     verify_self_certification,
 };
-// SigningKeyId is defined in this module and exported directly.
+// SigningKeyId re-exported from scp-primitives (see pub use above).
 pub use dht_client::{DhtClient, InMemoryDhtClient};
 #[cfg(feature = "production-dht")]
 pub use dht_client::{PkarrDhtClient, PkarrDhtClientBuilder};
@@ -68,185 +68,11 @@ pub use resolver::{
 };
 
 use serde::{Deserialize, Serialize};
-use std::fmt;
 
 use scp_platform::traits::{KeyCustody, KeyHandle};
 
-// ---------------------------------------------------------------------------
-// DID newtype (SCP-187)
-// ---------------------------------------------------------------------------
-
-/// Decentralized Identifier string (e.g., `"did:dht:z6Mk..."`).
-///
-/// A newtype wrapper around `String` providing type safety across the SCP
-/// codebase. Replaces the independent `type DID = String` aliases that were
-/// previously scattered across modules.
-///
-/// Implements `Deref<Target = str>` for ergonomic access to `&str` methods,
-/// `Borrow<str>` for `HashMap`/`HashSet` lookups with `&str` keys, and
-/// `#[serde(transparent)]` for zero-overhead JSON serialization.
-///
-/// See SCP-187 in `.docs/prds/prd.json`.
-#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash, Serialize, Deserialize)]
-#[serde(transparent)]
-pub struct DID(pub String);
-
-impl std::ops::Deref for DID {
-    type Target = str;
-    fn deref(&self) -> &str {
-        &self.0
-    }
-}
-
-impl AsRef<str> for DID {
-    fn as_ref(&self) -> &str {
-        &self.0
-    }
-}
-
-impl std::fmt::Display for DID {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(&self.0)
-    }
-}
-
-impl From<String> for DID {
-    fn from(s: String) -> Self {
-        Self(s)
-    }
-}
-
-impl From<&str> for DID {
-    fn from(s: &str) -> Self {
-        Self(s.to_owned())
-    }
-}
-
-impl PartialEq<str> for DID {
-    fn eq(&self, other: &str) -> bool {
-        self.0 == other
-    }
-}
-
-impl PartialEq<&str> for DID {
-    fn eq(&self, other: &&str) -> bool {
-        self.0 == *other
-    }
-}
-
-impl PartialEq<String> for DID {
-    fn eq(&self, other: &String) -> bool {
-        self.0 == *other
-    }
-}
-
-impl std::borrow::Borrow<str> for DID {
-    fn borrow(&self) -> &str {
-        &self.0
-    }
-}
-
-// ---------------------------------------------------------------------------
-// SigningKeyId (ADR-039)
-// ---------------------------------------------------------------------------
-
-/// Identifies which verification method signed an action.
-///
-/// Used in `ScpCredential`, `InnerEnvelope`, and `SenderKeyEpochAdvance` to
-/// indicate whether the `#active` (human) or `#agent` (agent software) signing
-/// key produced a signature. Verifiers resolve the correct public key from the
-/// sender's DID document using this field.
-///
-/// Wire-serializes as `"#active"` / `"#agent"` for JSON interoperability.
-///
-/// See ADR-039 in `.docs/adrs/phase-1.md`.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub enum SigningKeyId {
-    /// The human's active signing key (`#active` verification method).
-    Active,
-    /// The agent's signing key (`#agent` verification method).
-    Agent,
-}
-
-impl SigningKeyId {
-    /// Returns the full DID document fragment reference (e.g., `"#active"` or `"#agent"`).
-    ///
-    /// This is the canonical string representation used in serialization,
-    /// display, and hash preimages.
-    #[must_use]
-    pub const fn as_fragment(&self) -> &'static str {
-        match self {
-            Self::Active => "#active",
-            Self::Agent => "#agent",
-        }
-    }
-
-    /// Alias for [`as_fragment`](Self::as_fragment).
-    #[must_use]
-    pub const fn fragment_ref(&self) -> &'static str {
-        self.as_fragment()
-    }
-
-    /// Returns the bare fragment name without the `#` prefix (e.g., `"active"` or `"agent"`).
-    #[must_use]
-    pub const fn fragment(&self) -> &'static str {
-        match self {
-            Self::Active => "active",
-            Self::Agent => "agent",
-        }
-    }
-
-    /// Returns the canonical byte representation for inclusion in hash
-    /// preimages.
-    ///
-    /// This is the UTF-8 encoding of [`as_fragment`](Self::as_fragment).
-    #[must_use]
-    pub const fn as_bytes(&self) -> &'static [u8] {
-        match self {
-            Self::Active => b"#active",
-            Self::Agent => b"#agent",
-        }
-    }
-}
-
-impl Default for SigningKeyId {
-    /// Defaults to [`SigningKeyId::Active`] for backward compatibility with
-    /// envelopes and protocol messages created before agent binding (ADR-039).
-    fn default() -> Self {
-        Self::Active
-    }
-}
-
-impl fmt::Display for SigningKeyId {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.write_str(self.as_fragment())
-    }
-}
-
-impl Serialize for SigningKeyId {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        serializer.serialize_str(self.as_fragment())
-    }
-}
-
-impl<'de> Deserialize<'de> for SigningKeyId {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        let s = String::deserialize(deserializer)?;
-        match s.as_str() {
-            "#active" => Ok(Self::Active),
-            "#agent" => Ok(Self::Agent),
-            other => Err(serde::de::Error::custom(format!(
-                "unknown SigningKeyId: {other}, expected \"#active\" or \"#agent\""
-            ))),
-        }
-    }
-}
+// Re-export DID and SigningKeyId from scp-primitives for backward compatibility.
+pub use scp_primitives::{DID, SigningKeyId};
 
 /// An SCP identity containing the DID string, key handles, and pre-rotation
 /// commitment.
@@ -456,82 +282,4 @@ pub trait DidMethod: Send + Sync {
         identity: &ScpIdentity,
         key_custody: &impl KeyCustody,
     ) -> impl Future<Output = Result<(ScpIdentity, DidDocument), IdentityError>> + Send;
-}
-
-#[cfg(test)]
-#[allow(clippy::unwrap_used, clippy::expect_used, clippy::clone_on_copy)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn signing_key_id_fragment() {
-        assert_eq!(SigningKeyId::Active.fragment(), "active");
-        assert_eq!(SigningKeyId::Agent.fragment(), "agent");
-    }
-
-    #[test]
-    fn signing_key_id_fragment_ref() {
-        assert_eq!(SigningKeyId::Active.fragment_ref(), "#active");
-        assert_eq!(SigningKeyId::Agent.fragment_ref(), "#agent");
-    }
-
-    #[test]
-    fn signing_key_id_display() {
-        assert_eq!(format!("{}", SigningKeyId::Active), "#active");
-        assert_eq!(format!("{}", SigningKeyId::Agent), "#agent");
-    }
-
-    #[test]
-    fn signing_key_id_serialize() {
-        let active_json = serde_json::to_string(&SigningKeyId::Active).unwrap();
-        assert_eq!(active_json, "\"#active\"");
-
-        let agent_json = serde_json::to_string(&SigningKeyId::Agent).unwrap();
-        assert_eq!(agent_json, "\"#agent\"");
-    }
-
-    #[test]
-    fn signing_key_id_deserialize() {
-        let active: SigningKeyId = serde_json::from_str("\"#active\"").unwrap();
-        assert_eq!(active, SigningKeyId::Active);
-
-        let agent: SigningKeyId = serde_json::from_str("\"#agent\"").unwrap();
-        assert_eq!(agent, SigningKeyId::Agent);
-    }
-
-    #[test]
-    fn signing_key_id_deserialize_unknown() {
-        let result = serde_json::from_str::<SigningKeyId>("\"#unknown\"");
-        assert!(result.is_err());
-        let err = result.unwrap_err().to_string();
-        assert!(
-            err.contains("unknown SigningKeyId"),
-            "error should mention unknown SigningKeyId, got: {err}"
-        );
-    }
-
-    #[test]
-    fn signing_key_id_roundtrip() {
-        for key_id in [SigningKeyId::Active, SigningKeyId::Agent] {
-            let json = serde_json::to_string(&key_id).unwrap();
-            let parsed: SigningKeyId = serde_json::from_str(&json).unwrap();
-            assert_eq!(key_id, parsed);
-        }
-    }
-
-    #[test]
-    fn signing_key_id_equality() {
-        assert_eq!(SigningKeyId::Active, SigningKeyId::Active);
-        assert_eq!(SigningKeyId::Agent, SigningKeyId::Agent);
-        assert_ne!(SigningKeyId::Active, SigningKeyId::Agent);
-    }
-
-    #[test]
-    fn signing_key_id_copy_clone() {
-        let original = SigningKeyId::Agent;
-        let copied = original;
-        let cloned = original.clone();
-        assert_eq!(original, copied);
-        assert_eq!(original, cloned);
-    }
 }
